+++ distributors.cabal
@@ -43,12 +43,13 @@ library
       Control.Lens.PartialIso
       Control.Lens.Wither
       Data.Profunctor.Distributor
-      Data.Profunctor.Do.Monadic
-      Data.Profunctor.Do.Polyadic
       Data.Profunctor.Filtrator
       Data.Profunctor.Grammar
       Data.Profunctor.Monadic
+      Data.Profunctor.Monadic.Do
       Data.Profunctor.Monoidal
+      Data.Profunctor.Polyadic
+      Data.Profunctor.Polyadic.Do
   other-modules:
       Paths_distributors
   autogen-modules:
